backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
aiofiles==24.1.0
python-dotenv==1.1.1
demucs @ git+https://github.com/facebookresearch/demucs@e976d93ecc3865e5757426930257e200846a520a
torch==2.0.1
torchaudio==2.0.2
yt-dlp==2025.6.30
scipy==1.15.3
numpy==1.24.3
celery==5.5.3
redis==5.2.1
librosa==0.11.0
twilio==9.6.5
pyrubberband==0.3.2
psycopg2==2.9.10
